+++ src/routes/+page.server.ts
@@ -1,8 +1,9 @@
 import { redirect } from "@sveltejs/kit";
 import { resolve } from "$app/paths";
+import { getDisplayablePackages } from "./packages.remote";
 
-export async function load({ parent }) {
-	const { displayablePackages } = await parent();
+export async function load() {
+	const displayablePackages = await getDisplayablePackages();
 	const firstCategory = displayablePackages[0];
 	if (!firstCategory) redirect(307, resolve("/packages"));
 	const firstPackage = firstCategory.packages[0];
