+++ src/routes/all-package-releases.ts
@@ -1,7 +1,7 @@
 import type { PostHog } from "posthog-node";
 import { dwarn } from "$lib/logging";
 import { discoverer } from "$lib/server/package-discoverer";
-import { getPackageReleases } from "./package/releases";
+import { getPackageReleases } from "./releases";
 
 /**
  * Get all the repositories and releases for all the
