+++ src/routes/package/[...package]/SidePanel.svelte
