+++ src/routes/package/[...package]/rss.ts
@@ -10,7 +10,7 @@ import {
 } from "$lib/properties";
 import { discoverer } from "$lib/server/package-discoverer";
 import { ALL_SLUG } from "$lib/types";
-import { getPackageReleases, getPackagesReleases } from "../releases";
+import { getPackageReleases, getPackagesReleases } from "../../releases";
 
 const dottedAlphaNumRegex = /[A-z\d]+\.[A-z\d]+$/;
 /**
